\name{comoOdeCpp-package}
\alias{comoOdeCpp-package}
\alias{comoOdeCpp}
\docType{package}
\title{
\packageTitle{comoOdeCpp}
}
\description{
\packageDescription{comoOdeCpp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{comoOdeCpp}
\packageIndices{comoOdeCpp}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{comoOdeCpp}
Ricardo Aguas, Sompob Saralamba, Bo Gao

Maintainer: Bo Gao \packageMaintainer{comoOdeCpp}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
