\name{comoOdeCpp-package}
\alias{comoOdeCpp-package}
\alias{comoOdeCpp}
\docType{package}
\title{
\packageTitle{comoOdeCpp}
}
\description{
\packageDescription{comoOdeCpp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{comoOdeCpp}
\packageIndices{comoOdeCpp}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{comoOdeCpp}

Maintainer: \packageMaintainer{comoOdeCpp}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
